Refactored using of IdentityManager interface
