+++ app/src/main/java/co/smartreceipts/android/SmartReceiptsApplication.java
@@ -17,7 +17,6 @@ import co.smartreceipts.android.date.DateFormatter;
 import co.smartreceipts.android.di.AppComponent;
 import co.smartreceipts.android.di.BaseAppModule;
 import co.smartreceipts.android.di.DaggerAppComponent;
-import co.smartreceipts.android.identity.IdentityManager;
 import co.smartreceipts.android.images.PicassoInitializer;
 import co.smartreceipts.android.launch.OnLaunchDataPreFetcher;
 import co.smartreceipts.android.ocr.OcrManager;
@@ -36,6 +35,7 @@ import co.smartreceipts.android.utils.leaks.MemoryLeakMonitor;
 import co.smartreceipts.android.utils.rx.DefaultRxErrorHandler;
 import co.smartreceipts.android.versioning.AppVersionManager;
 import co.smartreceipts.aws.cognito.CognitoManager;
+import co.smartreceipts.core.identity.IdentityManager;
 import co.smartreceipts.core.utils.log.Logger;
 import dagger.android.AndroidInjector;
 import dagger.android.DispatchingAndroidInjector;
