+++ app/src/main/java/co/smartreceipts/android/ocr/widget/configuration/OcrConfigurationInteractor.kt
@@ -4,9 +4,6 @@ import co.smartreceipts.android.analytics.Analytics
 import co.smartreceipts.android.analytics.events.DataPoint
 import co.smartreceipts.android.analytics.events.DefaultDataPointEvent
 import co.smartreceipts.android.analytics.events.Events
-import co.smartreceipts.core.di.scopes.FragmentScope
-import co.smartreceipts.android.identity.IdentityManager
-import co.smartreceipts.core.identity.store.EmailAddress
 import co.smartreceipts.android.ocr.purchases.OcrPurchaseTracker
 import co.smartreceipts.android.purchases.PurchaseManager
 import co.smartreceipts.android.purchases.model.AvailablePurchase
@@ -16,6 +13,9 @@ import co.smartreceipts.android.purchases.model.PurchaseFamily
 import co.smartreceipts.android.purchases.source.PurchaseSource
 import co.smartreceipts.android.settings.UserPreferenceManager
 import co.smartreceipts.android.settings.catalog.UserPreference
+import co.smartreceipts.core.di.scopes.FragmentScope
+import co.smartreceipts.core.identity.IdentityManager
+import co.smartreceipts.core.identity.store.EmailAddress
 import co.smartreceipts.core.utils.log.Logger
 import io.reactivex.Observable
 import io.reactivex.Single
