+++ app/src/main/java/co/smartreceipts/android/purchases/subscriptions/RemoteSubscriptionManager.kt
@@ -1,10 +1,10 @@
 package co.smartreceipts.android.purchases.subscriptions
 
 import co.smartreceipts.android.apis.WebServiceManager
-import co.smartreceipts.android.identity.IdentityManager
 import co.smartreceipts.android.purchases.apis.subscriptions.SubscriptionsApiService
 import co.smartreceipts.android.purchases.wallet.PurchaseWallet
 import co.smartreceipts.core.di.scopes.ApplicationScope
+import co.smartreceipts.core.identity.IdentityManager
 import co.smartreceipts.core.utils.log.Logger
 import io.reactivex.Observable
 import io.reactivex.Scheduler
