+++ app/src/main/java/co/smartreceipts/android/push/PushManager.java
@@ -14,15 +14,14 @@ import javax.inject.Inject;
 import co.smartreceipts.android.analytics.Analytics;
 import co.smartreceipts.android.analytics.events.ErrorEvent;
 import co.smartreceipts.android.analytics.events.Events;
-import co.smartreceipts.core.di.scopes.ApplicationScope;
-import co.smartreceipts.android.identity.IdentityManager;
+import co.smartreceipts.android.identity.IdentityManagerImpl;
 import co.smartreceipts.android.push.apis.me.UpdatePushTokensRequest;
 import co.smartreceipts.android.push.apis.me.UpdateUserPushTokens;
-import co.smartreceipts.push.internal.FcmTokenRetriever;
-import co.smartreceipts.push.store.PushDataStore;
 import co.smartreceipts.core.di.scopes.ApplicationScope;
 import co.smartreceipts.core.utils.log.Logger;
 import co.smartreceipts.push.PushMessageReceiver;
+import co.smartreceipts.push.internal.FcmTokenRetriever;
+import co.smartreceipts.push.store.PushDataStore;
 import io.reactivex.Scheduler;
 import io.reactivex.schedulers.Schedulers;
 
@@ -30,7 +29,7 @@ import io.reactivex.schedulers.Schedulers;
 @ApplicationScope
 public class PushManager {
 
-    private final IdentityManager identityManager;
+    private final IdentityManagerImpl identityManager;
     private final Analytics analytics;
     private final FcmTokenRetriever fcmTokenRetriever;
     private final PushDataStore pushDataStore;
@@ -38,15 +37,14 @@ public class PushManager {
     private final CopyOnWriteArrayList<PushMessageReceiver> pushMessageReceivers = new CopyOnWriteArrayList<>();
 
     @Inject
-    public PushManager(@NonNull Context context,
-                       @NonNull IdentityManager identityManager,
+    public PushManager(@NonNull IdentityManagerImpl identityManager,
                        @NonNull Analytics analytics,
                        @NonNull PushDataStore pushDataStore) {
         this(identityManager, analytics, new FcmTokenRetriever(), pushDataStore, Schedulers.io());
     }
 
     @VisibleForTesting
-    public PushManager(@NonNull IdentityManager identityManager,
+    public PushManager(@NonNull IdentityManagerImpl identityManager,
                        @NonNull Analytics analytics,
                        @NonNull FcmTokenRetriever fcmTokenRetriever,
                        @NonNull PushDataStore pushDataStore,
