+++ app/src/test/java/co/smartreceipts/android/ocr/OcrManagerTest.java
@@ -15,13 +15,11 @@ import java.io.File;
 
 import co.smartreceipts.android.analytics.Analytics;
 import co.smartreceipts.android.apis.WebServiceManager;
-import co.smartreceipts.aws.s3.S3Manager;
 import co.smartreceipts.android.config.ConfigurationManager;
-import co.smartreceipts.android.identity.IdentityManager;
 import co.smartreceipts.android.ocr.apis.OcrService;
 import co.smartreceipts.android.ocr.apis.model.OcrResponse;
-import co.smartreceipts.android.ocr.apis.model.RecognitionResponse;
 import co.smartreceipts.android.ocr.apis.model.RecognitionRequest;
+import co.smartreceipts.android.ocr.apis.model.RecognitionResponse;
 import co.smartreceipts.android.ocr.purchases.OcrPurchaseTracker;
 import co.smartreceipts.android.ocr.push.OcrPushMessageReceiver;
 import co.smartreceipts.android.ocr.push.OcrPushMessageReceiverFactory;
@@ -30,6 +28,8 @@ import co.smartreceipts.android.push.PushManager;
 import co.smartreceipts.android.settings.UserPreferenceManager;
 import co.smartreceipts.android.settings.catalog.UserPreference;
 import co.smartreceipts.android.utils.ConfigurableResourceFeature;
+import co.smartreceipts.aws.s3.S3Manager;
+import co.smartreceipts.core.identity.IdentityManager;
 import io.reactivex.Observable;
 import io.reactivex.observers.TestObserver;
