+++ app/src/test/java/co/smartreceipts/android/ocr/purchases/OcrPurchaseTrackerTest.java
@@ -14,9 +14,6 @@ import java.util.Set;
 import co.smartreceipts.android.apis.SmartReceiptsApiErrorResponse;
 import co.smartreceipts.android.apis.SmartReceiptsApiException;
 import co.smartreceipts.android.apis.WebServiceManager;
-import co.smartreceipts.android.identity.IdentityManager;
-import co.smartreceipts.core.identity.apis.me.MeResponse;
-import co.smartreceipts.core.identity.apis.me.User;
 import co.smartreceipts.android.purchases.PurchaseManager;
 import co.smartreceipts.android.purchases.apis.purchases.MobileAppPurchasesService;
 import co.smartreceipts.android.purchases.apis.purchases.PurchaseRequest;
@@ -27,6 +24,9 @@ import co.smartreceipts.android.purchases.model.ManagedProduct;
 import co.smartreceipts.android.purchases.model.PurchaseFamily;
 import co.smartreceipts.android.purchases.source.PurchaseSource;
 import co.smartreceipts.android.purchases.wallet.PurchaseWallet;
+import co.smartreceipts.core.identity.IdentityManager;
+import co.smartreceipts.core.identity.apis.me.MeResponse;
+import co.smartreceipts.core.identity.apis.me.User;
 import io.reactivex.Completable;
 import io.reactivex.Observable;
 import io.reactivex.schedulers.Schedulers;
