+++ app/src/test/java/co/smartreceipts/android/ocr/widget/configuration/OcrConfigurationInteractorTest.java
@@ -15,8 +15,6 @@ import java.util.List;
 import java.util.Set;
 
 import co.smartreceipts.android.analytics.Analytics;
-import co.smartreceipts.android.identity.IdentityManager;
-import co.smartreceipts.core.identity.store.EmailAddress;
 import co.smartreceipts.android.ocr.purchases.OcrPurchaseTracker;
 import co.smartreceipts.android.purchases.PurchaseManager;
 import co.smartreceipts.android.purchases.model.AvailablePurchase;
@@ -24,6 +22,8 @@ import co.smartreceipts.android.purchases.model.InAppPurchase;
 import co.smartreceipts.android.purchases.source.PurchaseSource;
 import co.smartreceipts.android.settings.UserPreferenceManager;
 import co.smartreceipts.android.settings.catalog.UserPreference;
+import co.smartreceipts.core.identity.IdentityManager;
+import co.smartreceipts.core.identity.store.EmailAddress;
 import io.reactivex.Observable;
 import io.reactivex.observers.TestObserver;
 import io.reactivex.subjects.PublishSubject;
